fix a bug

+++ cpp/game/boardhistory.h
@@ -134,7 +134,7 @@ struct BoardHistory {
 
 private:
   bool koHashOccursInHistory(Hash128 koHash, const KoHashTable* rootKoHashTable) const;
-  int countAreaScoreWhiteMinusBlack(const Board& board, Color area[Board::MAX_ARR_SIZE]) const;
+  float countAreaScoreWhiteMinusBlack(const Board& board, Color area[Board::MAX_ARR_SIZE]) const;
   void setFinalScoreAndWinner(float score);
   int newConsecutiveEndingPassesAfterPass() const;
 };
